+++ Calculator/Gui/main.cpp
@@ -6,8 +6,8 @@ int main(int argc, char *argv[])
     QApplication app(argc, argv);
 
     MainWindow window;
-    window.setMinimumSize(520, 420);
-    window.resize(560, 460);
+    window.setMinimumSize(560, 560);
+    window.resize(620, 610);
     window.show();
 
     return app.exec();